Python/chatbot/requirements.txt
"fastapi[standard]"
google-generativeai
python-dotenv
markdown